+++ src/dependency_injector/providers.pyx
@@ -68,68 +68,60 @@ config_env_marker_pattern = re.compile(
     r'\${(?P<name>[^}^{:]+)(?P<separator>:?)(?P<default>.*?)}',
 )
 
-def _resolve_config_env_markers(config_value):
-    """"Replace environment variable markers with their values."""
-    for match in reversed(list(config_env_marker_pattern.finditer(config_value))):
+def _resolve_config_env_markers(config_content, envs_required=False):
+    """Replace environment variable markers with their values."""
+    findings = list(config_env_marker_pattern.finditer(config_content))
+
+    for match in reversed(findings):
+        env_name = match.group('name')
         has_default = match.group('separator') == ':'
 
-        value = os.getenv(match.group('name'))
+        value = os.getenv(env_name)
         if value is None:
-            if not has_default:
-                continue
+            if not has_default and envs_required:
+                raise ValueError(f'Missing required environment variable "{env_name}"')
             value = match.group('default')
 
         span_min, span_max = match.span()
-        config_value = f'{config_value[:span_min]}{value}{config_value[span_max:]}'
-    return config_value
+        config_content = f'{config_content[:span_min]}{value}{config_content[span_max:]}'
+    return config_content
 
 
 if sys.version_info[0] == 3:
-    class EnvInterpolation(iniconfigparser.BasicInterpolation):
-        """Interpolation which expands environment variables in values."""
-
-        def before_get(self, parser, section, option, value, defaults):
-            value = super().before_get(parser, section, option, value, defaults)
-            return _resolve_config_env_markers(value)
-
-    def _parse_ini_file(filepath):
-        parser = iniconfigparser.ConfigParser(interpolation=EnvInterpolation())
+    def _parse_ini_file(filepath, envs_required=False):
+        parser = iniconfigparser.ConfigParser()
         with open(filepath) as config_file:
-            parser.read_file(config_file)
+            config_string = _resolve_config_env_markers(
+                config_file.read(),
+                envs_required=envs_required,
+            )
+        parser.read_string(config_string)
         return parser
 else:
     import StringIO
 
-    def _parse_ini_file(filepath):
+    def _parse_ini_file(filepath, envs_required=False):
         parser = iniconfigparser.ConfigParser()
         with open(filepath) as config_file:
-            config_string = _resolve_config_env_markers(config_file.read())
+            config_string = _resolve_config_env_markers(
+                config_file.read(),
+                envs_required=envs_required,
+            )
         parser.readfp(StringIO.StringIO(config_string))
         return parser
 
 
 if yaml:
-    # TODO: use SafeLoader without env interpolation by default in version 5.*
-    def yaml_env_marker_constructor(_, node):
-        """"Replace environment variable marker with its value."""
-        return _resolve_config_env_markers(node.value)
-
-    yaml.add_implicit_resolver('!path', config_env_marker_pattern)
-    yaml.add_constructor('!path', yaml_env_marker_constructor)
-
     class YamlLoader(yaml.SafeLoader):
-        """Custom YAML loader.
+        """YAML loader.
 
-        Inherits ``yaml.SafeLoader`` and add environment variables interpolation.
+        This loader mimics ``yaml.SafeLoader``.
         """
-
-    YamlLoader.add_implicit_resolver('!path', config_env_marker_pattern, None)
-    YamlLoader.add_constructor('!path', yaml_env_marker_constructor)
 else:
     class YamlLoader:
-        """Custom YAML loader.
+        """YAML loader.
 
-        Inherits ``yaml.SafeLoader`` and add environment variables interpolation.
+        This loader mimics ``yaml.SafeLoader``.
         """
 
 
@@ -1535,7 +1527,7 @@ cdef class ConfigurationOption(Provider):
         """
         self.override(value)
 
-    def from_ini(self, filepath, required=UNDEFINED):
+    def from_ini(self, filepath, required=UNDEFINED, envs_required=False):
         """Load configuration from the ini file.
 
         Loaded configuration is merged recursively over existing configuration.
@@ -1546,10 +1538,16 @@ cdef class ConfigurationOption(Provider):
         :param required: When required is True, raise an exception if file does not exist.
         :type required: bool
 
+        :param envs_required: When True, raises an error on undefined environment variable.
+        :type envs_required: bool
+
         :rtype: None
         """
         try:
-            parser = _parse_ini_file(filepath)
+            parser = _parse_ini_file(
+                filepath,
+                envs_required=envs_required or self._is_strict_mode_enabled(),
+            )
         except IOError as exception:
             if required is not False \
                     and (self._is_strict_mode_enabled() or required is True) \
@@ -1567,7 +1565,7 @@ cdef class ConfigurationOption(Provider):
             current_config = {}
         self.override(merge_dicts(current_config, config))
 
-    def from_yaml(self, filepath, required=UNDEFINED, loader=None):
+    def from_yaml(self, filepath, required=UNDEFINED, loader=None, envs_required=False):
         """Load configuration from the yaml file.
 
         Loaded configuration is merged recursively over existing configuration.
@@ -1580,6 +1578,9 @@ cdef class ConfigurationOption(Provider):
 
         :param loader: YAML loader, :py:class:`YamlLoader` is used if not specified.
         :type loader: ``yaml.Loader``
+
+        :param envs_required: When True, raises an error on undefined environment variable.
+        :type envs_required: bool
 
         :rtype: None
         """
@@ -1595,7 +1596,7 @@ cdef class ConfigurationOption(Provider):
 
         try:
             with open(filepath) as opened_file:
-                config = yaml.load(opened_file, loader)
+                config_content = opened_file.read()
         except IOError as exception:
             if required is not False \
                     and (self._is_strict_mode_enabled() or required is True) \
@@ -1603,6 +1604,12 @@ cdef class ConfigurationOption(Provider):
                 exception.strerror = 'Unable to load configuration file {0}'.format(exception.strerror)
                 raise
             return
+
+        config_content = _resolve_config_env_markers(
+            config_content,
+            envs_required=envs_required or self._is_strict_mode_enabled(),
+        )
+        config = yaml.load(config_content, loader)
 
         current_config = self.__call__()
         if not current_config:
@@ -1956,7 +1963,7 @@ cdef class Configuration(Object):
         """
         self.override(value)
 
-    def from_ini(self, filepath, required=UNDEFINED):
+    def from_ini(self, filepath, required=UNDEFINED, envs_required=False):
         """Load configuration from the ini file.
 
         Loaded configuration is merged recursively over existing configuration.
@@ -1967,10 +1974,16 @@ cdef class Configuration(Object):
         :param required: When required is True, raise an exception if file does not exist.
         :type required: bool
 
+        :param envs_required: When True, raises an error on undefined environment variable.
+        :type envs_required: bool
+
         :rtype: None
         """
         try:
-            parser = _parse_ini_file(filepath)
+            parser = _parse_ini_file(
+                filepath,
+                envs_required=envs_required or self._is_strict_mode_enabled(),
+            )
         except IOError as exception:
             if required is not False \
                     and (self._is_strict_mode_enabled() or required is True) \
@@ -1988,7 +2001,7 @@ cdef class Configuration(Object):
             current_config = {}
         self.override(merge_dicts(current_config, config))
 
-    def from_yaml(self, filepath, required=UNDEFINED, loader=None):
+    def from_yaml(self, filepath, required=UNDEFINED, loader=None, envs_required=False):
         """Load configuration from the yaml file.
 
         Loaded configuration is merged recursively over existing configuration.
@@ -2001,6 +2014,9 @@ cdef class Configuration(Object):
 
         :param loader: YAML loader, :py:class:`YamlLoader` is used if not specified.
         :type loader: ``yaml.Loader``
+
+        :param envs_required: When True, raises an error on undefined environment variable.
+        :type envs_required: bool
 
         :rtype: None
         """
@@ -2016,7 +2032,7 @@ cdef class Configuration(Object):
 
         try:
             with open(filepath) as opened_file:
-                config = yaml.load(opened_file, loader)
+                config_content = opened_file.read()
         except IOError as exception:
             if required is not False \
                     and (self._is_strict_mode_enabled() or required is True) \
@@ -2024,6 +2040,12 @@ cdef class Configuration(Object):
                 exception.strerror = 'Unable to load configuration file {0}'.format(exception.strerror)
                 raise
             return
+
+        config_content = _resolve_config_env_markers(
+            config_content,
+            envs_required=envs_required or self._is_strict_mode_enabled(),
+        )
+        config = yaml.load(config_content, loader)
 
         current_config = self.__call__()
         if not current_config: